Modularize item needs table
